pyproject.toml
[tool.poetry]
name = "fastapi-template-app"
version = "0.1.0"
description = ""
authors = ["VladMerk <merkvladimir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic-settings = "^2.2.1"
aiosqlite = "^0.20.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
flake8-pyproject = "^1.2.3"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.flake8]
max-line-length = 100
extend-ignore = "W291 E128 E122 F401"

[tool.pytest.ini_options]
addopts = "-sv -p no:cacheprovider"
testpaths = ["tests"]

[tool.isort]
line_length = 100
multi_line_output = 3
